// frontend/src/generated/types/adminUser.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface AdminUser {
  sysId?: string;
  userId: string;
  userNm: string;
  passwordFailCnt?: string;
  initYn?: string;
  plantCd: string;
}