// frontend/src/generated/types/adminUserRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */
import type { AdminUser } from './adminUser';

export interface AdminUserRequest {
  newItems?: AdminUser[];
  updateItems?: AdminUser[];
  delItems?: AdminUser[];
}